add support for verifying downloads of terraform
This commit downloads the SHA256 checksum file for the version
and if there is an shasum binary available, uses that to test
the validity of the download.

If keybase is installed, it also leverages hashicorp's published
public key to validate that the checksum file is properly signed
by that published key.