prevent interactive prompting from keybase
This changes up the process a little to test to make sure
that the user is logged into keybase before trying to use
it. If logged in, it also verifies that they are following
hashicorp before trying to validate.

To call a bit more attention when key aspects are missing
it now highlights those relevant messages in yellow.