Document tersely GnuPG options; fix pygments
Don't document the full flexibility of the GnuPG enabling options, folks
who care enough can read the shell and see what the options are.  This
is power-user territory and if you need it, you should be reading the
code.

Adjust the code-block pygments declarations to be more accurate; instead
of `sh` for almost everything, use `console` or `puppet` as appropriate,
and add `terraform` to the relevant blocks.
<http://pygments.org/docs/lexers/> lists the common lexers.  Most
ecosystem tools copy whatever pygments does.