Support GnuPG tools, not just keybase, for OpenPGP signature verification (#109)
* Support using GnuPG gpg or gpgv to verify downloads

For power-users, optionally let them choose to use GnuPG's tools to
verify the download.  This requires that they have a path to the key, if
using GnuPG.

SECURITY: Also include a copy of the key for use by gpgv, but require
the caller to explicitly opt into trusting the key in our repo.  Trust
has to start somewhere, and they're calling scripts written here, but
avoid possible complaints by making it opt-in.

Key validated as identical (excluding comment) coming from
<https://keybase.io/hashicorp> and <https://www.hashicorp.com/security>;
matches what I had in my keyring too.  The form committed here is the
minimal version without any third-party signatures, made with:

    gpg --dearmor -o share/hashicorp-keys.pgp tmp-hc

The non-ASCII armored form is needed for `gpgv`.

nit: adjust existing `GPG` error messages to be `OpenPGP`:
 * OpenPGP is the standard
 * PGP is the original implementation, now a commercial company/unit
 * GnuPG is another implementation, now predominant
 * Keybase implements OpenPGP

* Document tersely GnuPG options; fix pygments

Don't document the full flexibility of the GnuPG enabling options, folks
who care enough can read the shell and see what the options are.  This
is power-user territory and if you need it, you should be reading the
code.

Adjust the code-block pygments declarations to be more accurate; instead
of `sh` for almost everything, use `console` or `puppet` as appropriate,
and add `terraform` to the relevant blocks.
<http://pygments.org/docs/lexers/> lists the common lexers.  Most
ecosystem tools copy whatever pygments does.