install: make keybase a fall-through verification variant
Prefer verification variants that require explicit opt-in by specifying
use-gnupg or use-gpgv. Use keybase as a fall-through verification
method in case nothing else has been explicitly requested.

In case a curated gpgv keyring is used for the trust, there is no point
in forcing to use keybase just because it is available while we
explicitly requested to use gpgv instead.