security: remove trust from revoked signing key
There is no point in trusting a revoked and exposed signing key, so lets
remove the trust from the vendored keyring. An exposed key must be
considered available to arbitrary attackers, hence it cannot reasonably
be still trusted.

Furthermore we explicitly download the shasum files that match the
current signing key. All old artifacts have been resigned in the
meanwhile. Even version 0.1.0 has the sums file for 72D7468F hence this
change is not a breaking change as long as we specifically use the new
sig file.

https://discuss.hashicorp.com/t/hcsec-2021-12-codecov-security-event-and-hashicorp-gpg-key-exposure/23512