Fix #301: find min-required version through -chdir
This commit makes it possible to use `min-required` with the terraform
configuration (including the `required_version` spec) in a subdirectory
as `terraform -chdir=subdir`.

Prior to this commit, this behaviour does not work (`-chdir` cannot be
used with `tfenv`'s `terraform` wrapper & `min-required` version)
because `tfenv` expects to find the configuration in `$PWD`.