Add `latest-allowed` option from required_version
This commit adds an option similar to `min-required` in that it parses
the terraform's `required_version`, but then uses the most recent
version allowed by that spec.

For example, given:

    required_version = "~> 0.10.0"

`min-required` would give us `0.10.0`, while `latest-allowed` will give
us `0.10.8` (via `latest:^0.10`).

The `<` operator is not implemented (only `>` `>=` `~>` `<=`) because
it's messy to find the latest version smaller than a given one in
`list-remote`'s output, given it may not (yet) exist, and it seems a
reasonably assumption/requirement that the version numbers will exist,
since, for example:

    required_version = "~> 1.0"

expresses the same as:

    required_version = "< 2.0.0"

without requiring a version that doesn't (at time of writing) exist.

This assumption/requirement stands for `<=` too, but again I think
that's reasonable since `<= 2.0.0` (as opposed to not equal to) seems
like a strange version constraint to use before it exists. (And when it
does, one can just use `2.0.0`.)

Closes #307.